pyproject.toml
[tool.poetry]
name = "bsort"
version = "0.1.0"
description = "Real-time bottle cap detection and sorting system for Raspberry Pi 5"
authors = ["ML Engineering Team"]
readme = "README.md"
packages = [{include = "bsort"}]

[tool.poetry.dependencies]
python = "^3.11"
# Core ML/CV dependencies
ultralytics = "^8.3.0"  # YOLO11 support
opencv-python-headless = "^4.10.0"  # Headless for ARM/Docker
numpy = "^1.26.0"
pillow = "^10.0.0"
pyyaml = "^6.0"

# ONNX Runtime for optimized inference
onnxruntime = "^1.19.0"  # CPU inference with XNNPACK/NEON optimization
onnx = "^1.16.0"

# MLOps and experiment tracking
wandb = "^0.18.0"

# CLI framework
typer = {extras = ["all"], version = "^0.12.0"}
rich = "^13.7.0"  # Beautiful terminal output

# Camera support for Raspberry Pi
picamera2 = {version = "^0.3.0", optional = true}

# Utilities
tqdm = "^4.66.0"
scikit-learn = "^1.5.0"

[tool.poetry.group.dev.dependencies]
# Code quality
pytest = "^8.0.0"
pytest-cov = "^5.0.0"
black = "^24.0.0"
isort = "^5.13.0"
pylint = "^3.2.0"
mypy = "^1.11.0"

# Jupyter for Task 1
jupyter = "^1.0.0"
ipykernel = "^6.29.0"
matplotlib = "^3.9.0"

[tool.poetry.extras]
rpi = ["picamera2"]

[tool.poetry.scripts]
bsort = "bsort.cli:app"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.pylint.messages_control]
max-line-length = 100
disable = [
    "C0103",  # Invalid name (for ML variables like X, y)
    "R0913",  # Too many arguments
    "R0914",  # Too many local variables
    "W0212",  # Access to protected member (needed for some ML libraries)
]

[tool.pylint.design]
max-args = 8
max-attributes = 12

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --cov=bsort --cov-report=term-missing --cov-report=html"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
